+++ components/src/main/java/com/mercadolibre/android/andesui/checkbox/align/AndesCheckboxAlignInterface.kt
@@ -1,0 +1,7 @@
+package com.mercadolibre.android.andesui.checkbox.align
+
+internal sealed class AndesCheckboxAlignInterface
+
+internal object AndesCheckboxLeft : AndesCheckboxAlignInterface()
+
+internal object AndesCheckboxRight : AndesCheckboxAlignInterface()