+++ components/src/main/java/com/mercadolibre/android/andesui/radiobutton/status/AndesRadioButtonStatus.kt
@@ -1,0 +1,19 @@
+package com.mercadolibre.android.andesui.radiobutton.status
+
+enum class AndesRadioButtonStatus {
+    SELECTED,
+    UNSELECTED;
+
+    companion object {
+        fun fromString(value: String): AndesRadioButtonStatus = valueOf(value.toUpperCase())
+    }
+
+    internal val status get() = getAndesRadioButtonStatus()
+
+    private fun getAndesRadioButtonStatus(): AndesRadioButtonStatusInterface {
+        return when (this) {
+            SELECTED -> AndesRadioButtonStatusSelected
+            UNSELECTED -> AndesRadioButtonStatusUnselected
+        }
+    }
+}