+++ demoapp/src/main/java/com/mercadolibre/android/andesui/demoapp/AndesSpecs.kt
@@ -9,12 +9,13 @@ private const val HOME = "https://company-161429.frontify.com/d/kxHCRixezmfK"
 private const val COMPONENT_PREFIX = "$HOME/n-a#/componentes"
 
 internal enum class AndesSpecs(val link: String) {
-
     HOME_PAGE(HOME),
+    CHECKBOX("$COMPONENT_PREFIX/checkbox"),
     TAG("$COMPONENT_PREFIX/tag"),
     BADGE("$COMPONENT_PREFIX/badge"),
     BUTTON("$COMPONENT_PREFIX/button"),
-    MESSAGE("$COMPONENT_PREFIX/message");
+    MESSAGE("$COMPONENT_PREFIX/message"),
+    RADIOBUTTON("$COMPONENT_PREFIX/radiobutton");
 }
 
 internal fun launchSpecs(context: Context, specs: AndesSpecs) {