+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v2.6.0
+## Added
+- Checkbox Component.
+- RadioButton Component.
+
 # v2.5.1
 ## Added
 - Add TextWatcher in TextField and TextArea.